+++ src/Integral_xPow2_Sim.java
@@ -1,12 +1,12 @@
 class Integral_xPow2_Sim extends IntegralGivenRectSim {
 
-    Integral_xPow2_Sim(float minX, float maxX, float minY, float maxY) {
+    Integral_xPow2_Sim(double minX, double maxX, double minY, double maxY) {
         super(minX, maxX, minY, maxY);
         resultCoefficient = 1.0f;
     }
 
     @Override
-    boolean isContainedPoint(float x, float y) {
+    boolean isContainedPoint(double x, double y) {
         return (y <= x * x);
     }
 }
